split scatters
