+++ tasks/mpi/guseynov_e_boost_scatter/include/ops_mpi.hpp
@@ -1,0 +1,43 @@
+#pragma once
+
+#include <gtest/gtest.h>
+
+#include <boost/mpi/collectives.hpp>
+#include <boost/mpi/communicator.hpp>
+#include <cmath>
+#include <numbers>
+#include <vector>
+
+#include "core/task/include/task.hpp"
+
+namespace guseynov_e_boost_scatter_mpi {
+
+std::vector<int> getRandomVector(int sz);
+
+class TestMPITaskSequential : public ppc::core::Task {
+ public:
+  explicit TestMPITaskSequential(std::shared_ptr<ppc::core::TaskData> taskData_) : Task(std::move(taskData_)) {}
+  bool pre_processing() override;
+  bool validation() override;
+  bool run() override;
+  bool post_processing() override;
+
+ private:
+  std::vector<int> input_;
+  int res_{};
+};
+
+class TestMPITaskParallel : public ppc::core::Task {
+ public:
+  explicit TestMPITaskParallel(std::shared_ptr<ppc::core::TaskData> taskData_) : Task(std::move(taskData_)) {}
+  bool pre_processing() override;
+  bool validation() override;
+  bool run() override;
+  bool post_processing() override;
+
+ private:
+  std::vector<int> input_, local_input_;
+  int res_{};
+  boost::mpi::communicator world;
+};
+}  // namespace guseynov_e_boost_scatter_mpi
