+++ tasks/mpi/guseynov_e_my_scatter/include/ops_mpi.hpp
@@ -10,7 +10,7 @@
 
 #include "core/task/include/task.hpp"
 
-namespace guseynov_e_scatter_mpi {
+namespace guseynov_e_my_scatter_mpi {
 
 std::vector<int> getRandomVector(int sz);
 
@@ -25,20 +25,6 @@ class TestMPITaskSequential : public ppc::core::Task {
  private:
   std::vector<int> input_;
   int res_{};
-};
-
-class TestMPITaskParallel : public ppc::core::Task {
- public:
-  explicit TestMPITaskParallel(std::shared_ptr<ppc::core::TaskData> taskData_) : Task(std::move(taskData_)) {}
-  bool pre_processing() override;
-  bool validation() override;
-  bool run() override;
-  bool post_processing() override;
-
- private:
-  std::vector<int> input_, local_input_;
-  int res_{};
-  boost::mpi::communicator world;
 };
 
 class MyScatterTestMPITaskParallel : public ppc::core::Task {
@@ -88,4 +74,4 @@ class MyScatterTestMPITaskParallel : public ppc::core::Task {
   int res_{};
   boost::mpi::communicator world;
 };
-}  // namespace guseynov_e_scatter_mpi
+}  // namespace guseynov_e_my_scatter_mpi
